Optimized class to be more generic
If we want to take a next step here, it's probably to pull fontFamily from application theme. Since there is no view to pull fontFamily from it's attributes.
We can also write a Congif model or something that takes care of font size, text color, direction and the typeface, then get rid of most of these overrides.
Thoughts?